moves components into component files, switches to , moves data into data file
